fix attachment delete via API

+++ classes/api/ApiAttachmentDelete.php
@@ -20,16 +20,29 @@
     public function run() {
       if (File::exists($this->data['id'])) {
 
-        # remove article_attachments
-        $conds  = array('attachment_id = ?', 'i', array($this->data['id']));
-        $res    = $this->db->delete('article_attachments', $conds);
-
-        # remove attachment
+        # get file path
+        $fields = array('file_path');
         $conds  = array('id = ?', 'i', array($this->data['id']));
-        $res    = $this->db->delete('attachments', $conds);
+        $res    = $this->db->select('attachments', $fields, $conds);
 
-        if ($res) {
-          echo 'success';
+        # delete file
+        $deleted = File::delete($res[0]['file_path']);
+
+        if ($deleted) {
+          # remove article_attachments
+          $conds  = array('attachment_id = ?', 'i', array($this->data['id']));
+          $res    = $this->db->delete('article_attachments', $conds);
+
+          # remove attachment
+          $conds  = array('id = ?', 'i', array($this->data['id']));
+          $res    = $this->db->delete('attachments', $conds);
+
+          if ($res) {
+            echo 'success';
+
+          } else {
+            echo 'error';
+          }
 
         } else {
           echo 'error';
